+++ linux/src/input/mod.rs
@@ -1,0 +1,5 @@
+mod gamepad;
+mod mouse;
+
+pub use gamepad::init_gamepad;
+pub use mouse::init_mouse;