+++ linux/src/main.rs
@@ -1,0 +1,5 @@
+use stickdeck_linux;
+
+fn main() -> anyhow::Result<()> {
+	stickdeck_linux::main()
+}